Fix issue with http interceptor error handler that would call authService logout function when a 401 was returned if a user entered invalid user/password.
In the event that the current request is to the login endpoint
we skip the standard error handler and simply return the error
as an observable.

This allows login component to subscribe to it and decorate the
UI with a warning when an invalid user/pass was supplied rather
than being refreshed by the logout action and not showing an
error.