Update dynamic environment and add new properties.
Change get() for dynamic environment from 'environment' to 'dynamic'. This makes it easier to understand when variables are being accessed from the dynamic environment or the static environment.

Remove unused variables from static environment.

Update wikiBase to point to confluence and add snowBase to point to snow for tickets.

Update helptext that was IPv4 specific to include IPv6 as well.