bug/generated client fix (#642)
* Add rimraf and use it to clear generated client before regenerating a new one.

This allows us to remove parts of the generated client that have been removed from the spec.

The removed client files in this commit demonstrate the fix.

* Safely enumerate messages on the dashboard.

The earlier method would result in the dashboard breaking and the navbar not displaying if the messages request failed. Now its handled more gracefully.